working copy. Needs to have some error protection added to userAddJet and possible other places JetsApplication 1) working load Cargo and dog fight AIRFIELD 1) setJet refactored to add Cargo and Fighter to correct classes so when add Cargo or fight is called new planes show up in display.

+++ src/com/skilldistillery/jets/entities/BasicPlane.java
@@ -2,7 +2,7 @@ package com.skilldistillery.jets.entities;
 
 public class BasicPlane extends Jet	{
 
-	public BasicPlane(String type, String model, int speed, int range, double price) {
+	public BasicPlane(String type, String model, int speed, int range, long price) {
 		super(type, model, speed, range, price);
 	}
